Bugfix: Switching out or switching against a Pokémon with max HP below 4 freezes the game
This happens because switching involves calculating a percentage of maximum enemy HP. Directly calculating HP × 100 / max HP would require a two-byte denominator, so instead the game calculates HP × 25 / (max HP / 4), since even a maximum HP of 999 divided by 4 is 249, which fits in one byte. However, if the maximum HP is below 4 this will divide by 0, which enters an infinite loop in _Divide.